+++ common/src/main/java/earth/terrarium/botarium/common/fluid/impl/WrappedBlockFluidContainer.java
@@ -81,9 +81,12 @@ public record WrappedBlockFluidContainer(BlockEntity block,
 
     @Override
     public long extractFromSlot(FluidHolder fluidHolder, FluidHolder toInsert, Runnable snapshot) {
-        long extract = container.extractFromSlot(fluidHolder, toInsert, snapshot);
-        update();
-        return extract;
+        return container.extractFromSlot(fluidHolder, toInsert, snapshot);
+    }
+
+    @Override
+    public long extractFromSlot(int slot, FluidHolder toExtract, boolean simulate) {
+        return container.extractFromSlot(slot, toExtract, simulate);
     }
 
     @Override