+++ fabric/src/main/java/earth/terrarium/botarium/fabric/fluid/storage/FabricBlockFluidContainer.java
@@ -5,20 +5,18 @@ import earth.terrarium.botarium.common.fluid.base.FluidHolder;
 import earth.terrarium.botarium.common.fluid.base.FluidSnapshot;
 import earth.terrarium.botarium.fabric.fluid.holder.FabricFluidHolder;
 import earth.terrarium.botarium.fabric.fluid.holder.ManualSyncing;
-import earth.terrarium.botarium.fabric.fluid.holder.WrappedFluidHolder;
 import earth.terrarium.botarium.util.Updatable;
 import net.fabricmc.fabric.api.transfer.v1.fluid.FluidVariant;
 import net.fabricmc.fabric.api.transfer.v1.storage.Storage;
 import net.fabricmc.fabric.api.transfer.v1.storage.StorageView;
 import net.fabricmc.fabric.api.transfer.v1.transaction.TransactionContext;
-import net.minecraft.world.level.block.entity.BlockEntity;
 
 import java.util.Iterator;
 import java.util.List;
 import java.util.stream.IntStream;
 
 public class FabricBlockFluidContainer<T extends FluidContainer & Updatable> extends ExtendedFluidContainer implements Storage<FluidVariant>, ManualSyncing {
-    private final T container;
+    protected final T container;
 
     public FabricBlockFluidContainer(T container) {
         this.container = container;
@@ -38,8 +36,7 @@ public class FabricBlockFluidContainer<T extends FluidContainer & Updatable> ext
 
     @Override
     public Iterator<StorageView<FluidVariant>> iterator() {
-        List<FluidHolder> fluids = container.getFluids();
-        return IntStream.range(0, fluids.size()).mapToObj(index -> new WrappedFluidHolder(this, fluids.get(index), container::extractFromSlot, container.getTankCapacity(index))).map(holder -> (StorageView<FluidVariant>) holder).iterator();
+        return IntStream.range(0, container.getSize()).mapToObj(index -> new SingleFluidSlot(this, index)).map(holder -> (StorageView<FluidVariant>) holder).iterator();
     }
 
     @Override