+++ fabric/src/main/java/earth/terrarium/botarium/fabric/fluid/storage/FabricItemFluidContainer.java
@@ -2,10 +2,7 @@ package earth.terrarium.botarium.fabric.fluid.storage;
 
 import earth.terrarium.botarium.common.fluid.base.FluidContainer;
 import earth.terrarium.botarium.common.fluid.base.FluidHolder;
-import earth.terrarium.botarium.common.fluid.base.FluidSnapshot;
 import earth.terrarium.botarium.fabric.fluid.holder.FabricFluidHolder;
-import earth.terrarium.botarium.fabric.fluid.holder.ManualSyncing;
-import earth.terrarium.botarium.fabric.fluid.holder.WrappedFluidHolder;
 import net.fabricmc.fabric.api.transfer.v1.context.ContainerItemContext;
 import net.fabricmc.fabric.api.transfer.v1.fluid.FluidVariant;
 import net.fabricmc.fabric.api.transfer.v1.item.ItemVariant;
@@ -14,14 +11,15 @@ import net.fabricmc.fabric.api.transfer.v1.storage.StorageView;
 import net.fabricmc.fabric.api.transfer.v1.transaction.TransactionContext;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.world.item.ItemStack;
+import org.jetbrains.annotations.NotNull;
 
 import java.util.Iterator;
 import java.util.List;
 import java.util.stream.IntStream;
 
 @SuppressWarnings("UnstableApiUsage")
-public class FabricItemFluidContainer extends ExtendedFluidContainer implements Storage<FluidVariant>, ManualSyncing {
-    private final FluidContainer container;
+public class FabricItemFluidContainer implements Storage<FluidVariant> {
+    protected final FluidContainer container;
     private final ContainerItemContext ctx;
 
     public FabricItemFluidContainer(ContainerItemContext ctx, FluidContainer container) {
@@ -33,7 +31,6 @@ public class FabricItemFluidContainer extends ExtendedFluidContainer implements 
 
     @Override
     public long insert(FluidVariant resource, long maxAmount, TransactionContext transaction) {
-        updateSnapshots(transaction);
         long inserted = container.insertFluid(FabricFluidHolder.of(resource, maxAmount), false);
         setChanged(transaction);
         return inserted;
@@ -41,33 +38,16 @@ public class FabricItemFluidContainer extends ExtendedFluidContainer implements 
 
     @Override
     public long extract(FluidVariant resource, long maxAmount, TransactionContext transaction) {
-        updateSnapshots(transaction);
         long extracted = container.extractFluid(FabricFluidHolder.of(resource, maxAmount), false).getFluidAmount();
         setChanged(transaction);
         return extracted;
     }
 
     @Override
-    public Iterator<StorageView<FluidVariant>> iterator() {
-        List<FluidHolder> fluids = container.getFluids();
-        return IntStream.range(0, fluids.size()).mapToObj(index -> new WrappedFluidHolder(this, fluids.get(index), container::extractFromSlot, container.getTankCapacity(index))).map(holder -> (StorageView<FluidVariant>) holder).iterator();
+    public @NotNull Iterator<StorageView<FluidVariant>> iterator() {
+        return IntStream.range(0, container.getSize()).mapToObj(index -> new SingleItemFluidSlot(this, index)).map(holder -> (StorageView<FluidVariant>) holder).iterator();
     }
 
-    @Override
-    public void onFinalCommit() {
-    }
-
-    @Override
-    public FluidSnapshot createSnapshot() {
-        return container.createSnapshot();
-    }
-
-    @Override
-    public void readSnapshot(FluidSnapshot snapshot) {
-        container.readSnapshot(snapshot);
-    }
-
-    @Override
     public void setChanged(TransactionContext transaction) {
         ItemStack stack = ctx.getItemVariant().toStack();
         this.container.serialize(stack.getOrCreateTag());