+++ changelog.md
@@ -1,3 +1,4 @@
-# 3.0.7
+# 3.0.8
 
-- New feature: Fluid particles!
+- Fix fluid dupe bug on items
+- New extractFromSlot method to replace old one.